# packages/shared/generated/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intercal API",
    "description": "Intercal public contract — the single source of truth.\nTypeSpec compiles to OpenAPI 3.1 + JSON Schema; TS types and Pydantic models are generated\nfrom those. Do not hand-edit generated artifacts.",
    "version": "0.0.0"
  },
...
  "paths": {
    "/v1/claims/verify": {
      "get": {
        "operationId": "verifyClaim",
        "description": "Verify a claim against recorded evidence as of a date.",
        "parameters": [
          {
            "$ref": "#/components/parameters/VerifyClaimQuery.claim_text"
          },
          {
            "$ref": "#/components/parameters/VerifyClaimQuery.as_of_date"
          },
          {
            "$ref": "#/components/parameters/VerifyClaimQuery.token_budget"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimVerificationResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/delta": {
      "get": {
        "operationId": "getDelta",
        "description": "What changed about a topic since a date, with evidence, confidence, and a token-budgeted summary.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DeltaQuery.topic"
          },
          {
            "$ref": "#/components/parameters/DeltaQuery.since_date"
          },
          {
            "$ref": "#/components/parameters/DeltaQuery.token_budget"
          },
          {
            "$ref": "#/components/parameters/DeltaQuery.until_date"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/entity": {
      "get": {
        "operationId": "getEntity",
        "description": "Entity state, relationships, and fact history, optionally at a point in time.",
        "parameters": [
          {
            "$ref": "#/components/parameters/EntityQuery.name_or_id"
          },
          {
            "$ref": "#/components/parameters/EntityQuery.at_date"
          },
          {
            "$ref": "#/components/parameters/EntityQuery.token_budget"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/evidence": {
      "get": {
        "operationId": "searchEvidence",
        "description": "Search source-grounded evidence by query, date range, and sources.",
        "parameters": [
          {
            "$ref": "#/components/parameters/EvidenceQuery.query"
          },
          {
            "$ref": "#/components/parameters/EvidenceQuery.from_date"
          },
          {
            "$ref": "#/components/parameters/EvidenceQuery.to_date"
          },
          {
            "$ref": "#/components/parameters/EvidenceQuery.limit"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/feedback": {
      "post": {
        "operationId": "submitFeedback",
        "description": "Submit bounded feedback for operator review. Creates a review record only; canonical graph state is unchanged.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "operations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        }
      }
    },
    "/v1/freshness": {
      "get": {
        "operationId": "getFreshness",
        "description": "Report how fresh Intercal's knowledge is for a topic or entity.",
        "parameters": [
          {
            "$ref": "#/components/parameters/FreshnessQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreshnessReport"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/sources": {
      "get": {
        "operationId": "getSources",
        "description": "List the source documents backing an entity or claim.",
        "parameters": [
          {
            "$ref": "#/components/parameters/SourcesQuery.entity_or_claim_id"
          },
          {
            "$ref": "#/components/parameters/SourcesQuery.limit"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/subscriptions": {
      "get": {
        "operationId": "listSubscriptions",
        "description": "List active subscriptions owned by the authenticated API key.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionsResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ]
      },
      "post": {
        "operationId": "createSubscription",
        "description": "Create a polling or webhook subscription for a topic, entity, relationship type, or claim pattern.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          }
        }
      }
    },
    "/v1/subscriptions/delete": {
      "post": {
        "operationId": "deleteSubscription",
        "description": "Deactivate a subscription owned by the authenticated API key.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSubscriptionRequest"
              }
            }
          }
        }
      }
    },
    "/v1/subscriptions/dispatch": {
      "post": {
        "operationId": "dispatchSubscriptionNotifications",
        "description": "Enqueue bounded notifications for matching subscriptions after a known knowledge change.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchSubscriptionResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchSubscriptionRequest"
              }
            }
          }
        }
      }
    },
    "/v1/subscriptions/poll": {
      "post": {
        "operationId": "pollSubscriptionNotifications",
        "description": "Poll pending notifications for a subscription. Returned notifications are marked delivered.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionNotificationsResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollSubscriptionRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "DeltaQuery.since_date": {
        "name": "since_date",
        "in": "query",
        "required": true,
        "description": "Return changes recorded since this date (e.g. a model cutoff).",
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "DeltaQuery.token_budget": {
        "name": "token_budget",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "explode": false
      },
      "DeltaQuery.topic": {
        "name": "topic",
        "in": "query",
        "required": true,
        "description": "Topic, entity, or claim subject to compute changes for.",
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "DeltaQuery.until_date": {
        "name": "until_date",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "EntityQuery.at_date": {
        "name": "at_date",
        "in": "query",
        "required": false,
        "description": "Point-in-time read: entity state as of this date.",
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "EntityQuery.name_or_id": {
        "name": "name_or_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "EntityQuery.token_budget": {
        "name": "token_budget",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "explode": false
      },
      "EvidenceQuery.from_date": {
        "name": "from_date",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "EvidenceQuery.limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100
        },
        "explode": false
      },
      "EvidenceQuery.query": {
        "name": "query",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "EvidenceQuery.to_date": {
        "name": "to_date",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "FreshnessQuery": {
        "name": "topic_or_entity",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "SourcesQuery.entity_or_claim_id": {
        "name": "entity_or_claim_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "SourcesQuery.limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100
        },
        "explode": false
      },
      "VerifyClaimQuery.as_of_date": {
        "name": "as_of_date",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "explode": false
      },
      "VerifyClaimQuery.claim_text": {
        "name": "claim_text",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "explode": false
      },
      "VerifyClaimQuery.token_budget": {
        "name": "token_budget",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "explode": false
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        },
        "description": "A structured API/tool error."
      },
...
      "ClaimVerificationResponse": {
        "type": "object",
        "required": [
          "claimText",
          "verdict",
          "confidence",
          "supportingEvidence",
          "contradictingEvidence"
        ],
        "properties": {
          "claimText": {
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/VerificationVerdict"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "supportingEvidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Citation"
            }
          },
          "contradictingEvidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Citation"
            }
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DeltaResponse": {
        "type": "object",
        "required": [
          "topic",
          "since",
          "summary",
          "changedClaims",
          "changedEntities",
          "confidence",
          "freshness"
        ],
        "properties": {
          "topic": {
            "type": "string"
          },
          "since": {
            "type": "string",
            "format": "date-time"
          },
          "until": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/Digest"
          },
          "changedClaims": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Claim"
            }
          },
          "changedEntities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntitySummary"
            }
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessReport"
          }
        }
      },
...
      "EntityResponse": {
        "type": "object",
        "required": [
          "entity",
          "freshness"
        ],
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relationship"
            }
          },
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Claim"
            }
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessReport"
          }
        }
      },
...
      "EvidenceResponse": {
        "type": "object",
        "required": [
          "hits",
          "total"
        ],
        "properties": {
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceHit"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FeedbackRequest": {
        "type": "object",
        "required": [
          "targetType",
          "targetId",
          "concernType",
          "summary"
        ],
        "properties": {
          "targetType": {
            "$ref": "#/components/schemas/FeedbackTargetType"
          },
          "targetId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "concernType": {
            "$ref": "#/components/schemas/FeedbackConcernType"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          },
          "details": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "description": "Bounded public feedback that creates a review record without mutating canonical knowledge."
      },
      "FeedbackResponse": {
        "type": "object",
        "required": [
          "review"
        ],
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewRecord"
          }
        }
      },
...
      "FreshnessReport": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "type": "string"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "lastIngestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "coverage": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 1,
            "description": "Fraction of expected sources currently covered, in [0,1]."
          },
          "staleness": {
            "type": "string"
          }
        },
        "description": "How fresh Intercal's knowledge is for a target."
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development",
      "variables": {}
    }
  ]
}